clean up temporary stream
