+++ src/DocumentFormat.OpenXml.Framework/Packaging/PackageUriHandlingExtensions.cs
@@ -34,7 +34,7 @@ internal static class PackageUriHandlingExtensions
 
         if (feature.Package.FileOpenAccess == FileAccess.Read)
         {
-            if (ReplaceStream(features))
+            if (features.EnableWriteableStream())
             {
                 // Don't register for disposal as we don't want it to write out to the temporary stream
                 features.Set<IPackageFeature>(new MalformedUriHandlingPackage(feature));
@@ -49,49 +49,6 @@ internal static class PackageUriHandlingExtensions
         }
 
         return features;
-    }
-
-    /// <summary>
-    /// Attempts to replace the underlying stream so that we can modify it even in readonly situations via a copy
-    /// </summary>
-    private static bool ReplaceStream(IFeatureCollection features)
-    {
-        if (features.Get<IPackageStreamFeature>() is { Stream.CanWrite: false } feature &&
-            features.Get<IPackageFeature>() is { } packageFeature && packageFeature.Capabilities.HasFlagFast(PackageCapabilities.Reload))
-        {
-            var tempStream = new TemporaryFile();
-
-            features.GetRequired<IDisposableFeature>().Register(tempStream);
-
-            feature.Stream.Position = 0;
-            feature.Stream.CopyTo(tempStream.Stream);
-
-            tempStream.Stream.Position = 0;
-            feature.Stream = tempStream.Stream;
-
-            return true;
-        }
-
-        return false;
-    }
-
-    private sealed class TemporaryFile : IDisposable
-    {
-        private readonly string _path;
-
-        public TemporaryFile()
-        {
-            _path = Path.GetTempFileName();
-            Stream = File.Create(_path);
-        }
-
-        public Stream Stream { get; }
-
-        public void Dispose()
-        {
-            Stream.Dispose();
-            File.Delete(_path);
-        }
     }
 
     private static XDocument? WalkRelationships(IPackagePart part, KnownUris uris)
